chore: integrate main view and ticket view
add: add components
