+++ Baseball/Baseball/Source/Utils/Component/VisualEffectView.swift
@@ -1,0 +1,20 @@
+//
+//  VisualEffectView.swift
+//  Baseball
+//
+//  Created by Nayeon Kim on 7/6/24.
+//
+
+import SwiftUI
+
+struct VisualEffectView: UIViewRepresentable {
+    var effect: UIVisualEffect?
+    
+    func makeUIView(context: UIViewRepresentableContext<Self>) -> UIVisualEffectView {
+        return UIVisualEffectView()
+    }
+    
+    func updateUIView(_ uiView: UIVisualEffectView, context: UIViewRepresentableContext<Self>) {
+        uiView.effect = effect
+    }
+}
